feat: renamed `signalprocessing._interp_cubic_spline` to `signalprocessin.interpspline`
